requirement.txt
Flask == 2.3.2
flask-sqlalchemy == 3.0.5
Flask-Migrate == 4.0.4
Flask-JWT-Extended == 4.5.2

